feat: pagamento parcial com status 'partial'
- Backend: pay-transaction aceita paidAmount parcial, acumula no valuePaid
- Backend: novo status 'partial' em transactionOccurrences
- Backend: list-transactions e dashboard incluem valuePaid e tratam partial
- Backend: relatórios WhatsApp mostram restante de pagamentos parciais
- Backend: fetchOverdueTransactionsForAlerts inclui status partial
- Frontend: dialog de pagamento com CurrencyInput para valor parcial
- Frontend: tabela mostra barra de progresso e valor pago para partial
- Frontend: ícone/modal/alertas com estilo amber para status partial
- Frontend: UpcomingAlerts e OverdueTransactionsCard destacam parciais
- Frontend: status-distribution-chart com fatia Parcial
- Tipos gerados: todos status enums incluem 'partial'

+++ api/src/http/controllers/transaction/pay-transaction.controller.ts
@@ -19,7 +19,7 @@ export async function payTransactionController(
   await payTransactionService({
     id,
     paidAt: paidAt ? new Date(paidAt) : undefined,
-    valuePaid: paidAmount ? BigInt(Math.round(paidAmount * 100)) : undefined,
+    valuePaid: paidAmount ? Math.round(paidAmount * 100) : undefined,
   })
   return reply.status(204).send()
 }
